backend/requirements.txt
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.10
mongoengine==0.15.0
pymongo==2.8
pytz==2017.3
six==1.11.0
django-rest-framework-mongoengine==3.3.1